Add files via upload
PakCortex is a modern web platform that provides a chat service powered by artificial intelligence (AI) for students. This platform uses the powerful technology of ChatGPT to answer students' questions and fulfill their educational needs. The goal of PakCortex is to assist students in learning and provide them with access to the best educational resources.